Add 'Scalers' to profiles:
A scaler is a repeatable field that can represent something using a
rating out of a number the user defines.

A user can add as many scalers as necessary, each scaler has a name and
a number / out of scale fields.

The repeating field helpers and JS can be used for the future repeatable
fields.

Migration required!